Move get_ffmpeg_url() from GUI to ffmpeg_path.py
